Make new-wallet work if no password and --from param
